Incremente en Vista "Add Existence" - Se agregaron providers de Existencias y Sedes. - Se modificaron los modelos 'cliente', 'detalle_venta', 'existencia' y 'venta' para descargar el Float que causaba conflictos y cambiarlo por un double. Además se cambió el 'DateTime' del modelo Existencia por un String para facilitar el manejo. - Se agregó la pantalla 'AddExistenceScreen' que permite añadir existencias a una sede en específico. - Se cambió ligeramente la pantalla ReadExistence, pero aún falta muchas mejoras. - Se agregaron los endpoints 'Productos (App)', 'Productos - Only Name (App)', 'Producto (NameById)', 'Existencias - Details (Precio + Cantidad)' y 'Sedes Only Name'. - Se aplicó el patrón Debouncer para retrasar la ejecución de un código de consulta. - Se agregaron los servicios para la Existencia y para las Sedes. - Se modificó el servicio del Producto para consultar nombres de producto por ID y para una consulta general de todos los nombres. - Se modificó el tema clásico. - Se agregó el ViewModel para las Existencias. - Se modificó el CustomDropDownButton para aceptar cualquier tipo de variable como valor inicial. - Se modificó el CustomTextField para ser más flexible.

+++ lib/models/detalle_venta.dart
@@ -1,13 +1,9 @@
-import 'dart:ffi';
-
-
-
 class DetalleVenta {
   int? idDetalleVenta;
   final int idExistencia;
   final int idVenta;
   final int cantidadVenta;
-  final Float precioVenta;
+  final double precioVenta;
 
   DetalleVenta({
     this.idDetalleVenta,
